pipelines/requirements.txt
# PySpark BigQuery Pipeline Dependencies

pyspark==3.5.0
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
